Split FS module
